backend/requirements.txt
Django==2.2
djangorestframework==3.9.3
psycopg2==2.7.6.1
django-filter
factory_boy
django-cors-headers